+++ unpythonic/syntax/autocurry.py
@@ -85,7 +85,7 @@ def _autocurry(block_body):
                 return tree
 
             hascurry = self.state.hascurry
-            if type(tree) is Call and not isx(tree.func, "ExpandedAutorefMarker"):
+            if type(tree) is Call:
                 if has_curry(tree):  # detect decorated lambda with manual curry
                     # the lambda inside the curry(...) is the next Lambda node we will descend into.
                     hascurry = True